Google distance API fixes, added googlemapsutils to lib
